third_party/WebKit/LayoutTests/http/tests/security/text-track-crossorigin-expected.txt

CONSOLE ERROR: Text track from origin 'http://localhost:8000' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Text track at 'http://localhost:8000/security/resources/captions-with-access-control-headers.php?count=0' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Text track from origin 'http://localhost:8000' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests loading cross-domain <track>.

Loading without Access-Control-Allow-Origin header, no "crossorigin" attribute on <video>
EVENT(error)
PASS: shouldLoad should be 'false' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '3' and is.


Loading without Access-Control-Allow-Origin header, setting video.crossorigin to "anonymous"
EVENT(error)
PASS: shouldLoad should be 'false' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '3' and is.


Loading with Access-Control-Allow-Origin header, leaving video.crossorigin as "anonymous"
EVENT(load)
PASS: shouldLoad should be 'true' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '2' and is.


Loading with Access-Control-Allow-Origin and Access-Control-Allow-Credentials headers, setting video.crossorigin to "use-credentials"
EVENT(load)
PASS: shouldLoad should be 'true' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '2' and is.


Loading without Access-Control-Allow-Origin header, with a redirect, no "crossorigin" attribute on <video>
EVENT(error)
PASS: shouldLoad should be 'false' and is.
PASS: event.target should be '[object HTMLTrackElement]' and is.
PASS: trackElement.readyState should be '3' and is.


END OF TEST
